requirements.txt
mysql.connector
requests
webexteamssdk